(Really!!) Completed RtpDump Reader and Writer (All Formats Except Short form 100%) Reverse Engineered bark.rtp and made my files match the output. The spec on the site is out of date and the origional author of the tools does not have time to look into why this is.
Added RtpDump Modificiation Test.

Added RtpDump Header Only Test.

Added RtpDump File Comparison Test.

Updated Utility-
Made GetBytesUnsafe for effiency.
Removed TicksPerMicroSecond constant.

Updated RtpPacket to allow only header and changed size of all fields to byte rather than int. Changed SequenceNumber to ushort.

Updated other classes to support changes as stated.

+++ Rtp/JpegFrame.cs
@@ -480,7 +480,7 @@ namespace Media.Rtp
                 RtpPacket currentPacket = new RtpPacket();
                 SynchronizationSourceIdentifier = currentPacket.SynchronizationSourceIdentifier = (ssrc ?? (uint)SynchronizationSourceIdentifier);
                 currentPacket.TimeStamp = (uint)(timeStamp ?? Utility.DateTimeToNtpTimestamp(DateTime.UtcNow));
-                currentPacket.SequenceNumber = (int)(sequenceNo ?? DateTime.UtcNow.Ticks);
+                currentPacket.SequenceNumber = (ushort)(sequenceNo ?? DateTime.UtcNow.Ticks);
                 currentPacket.PayloadType = JpegFrame.PayloadType;
 
                 //Where we are in the current packet
@@ -608,7 +608,7 @@ namespace Media.Rtp
                                 currentPacket = new RtpPacket(payloadSize)
                                 {
                                     TimeStamp = currentPacket.TimeStamp,
-                                    SequenceNumber = currentPacket.SequenceNumber + 1,
+                                    SequenceNumber = (ushort)(currentPacket.SequenceNumber + 1),
                                     SynchronizationSourceIdentifier = currentPacket.SynchronizationSourceIdentifier,
                                     PayloadType = JpegFrame.PayloadType
                                 };
